clientside encryption test refactoring
